+++ client/packages/invoices/src/InboundShipment/DetailView/modals/index.ts
@@ -1,4 +1,3 @@
 export * from './InboundLineEdit';
 export * from './InboundServiceLineEdit';
 export * from './TaxEdit';
-export * from './NewReturnItems';
