+++ client/packages/invoices/src/InboundShipment/api/api.ts
@@ -193,14 +193,6 @@ export const getInboundQueries = (sdk: Sdk, storeId: string) => ({
 
       throw new Error('Could not find invoice!');
     },
-    newSupplierReturnLines: async (lineIds: string[]) => {
-      const result = await sdk.newSupplierReturnLines({
-        inboundShipmentLineIds: lineIds,
-        storeId,
-      });
-
-      return result?.newSupplierReturn;
-    },
   },
   delete: async (invoices: InboundRowFragment[]): Promise<string[]> => {
     const result =
