+++ client/packages/invoices/src/Returns/api/api.ts
@@ -1,0 +1,14 @@
+import { Sdk } from './operations.generated';
+
+export const getReturnsQueries = (sdk: Sdk, storeId: string) => ({
+  get: {
+    newSupplierReturnLines: async (lineIds: string[]) => {
+      const result = await sdk.newSupplierReturnLines({
+        inboundShipmentLineIds: lineIds,
+        storeId,
+      });
+
+      return result?.newSupplierReturn;
+    },
+  },
+});
