+++ client/packages/invoices/src/Returns/api/hooks/index.ts
@@ -1,0 +1,11 @@
+import { Lines } from './line';
+import { Utils } from './utils';
+
+export const useReturns = {
+  lines: {
+    newReturnLines: Lines.useNewSupplierReturnLines,
+  },
+  utils: {
+    api: Utils.useReturnsApi,
+  },
+};
