+++ client/packages/invoices/src/Returns/api/hooks/line/index.ts
@@ -1,0 +1,5 @@
+import { useNewSupplierReturnLines } from './useNewSupplierReturnLines';
+
+export const Lines = {
+  useNewSupplierReturnLines,
+};
