+++ client/packages/invoices/src/Returns/api/hooks/utils/index.ts
@@ -1,0 +1,5 @@
+import { useReturnsApi } from './useReturnsApi';
+
+export const Utils = {
+  useReturnsApi,
+};
