+++ client/packages/invoices/src/Returns/api/hooks/utils/useReturnsApi.ts
@@ -1,0 +1,18 @@
+import { useAuthContext, useGql } from '@openmsupply-client/common';
+import { getReturnsQueries } from '../../api';
+import { getSdk } from '../../operations.generated';
+
+export const useReturnsApi = () => {
+  const { storeId } = useAuthContext();
+  const keys = {
+    base: () => ['returns'] as const,
+    count: () => [...keys.base(), 'count'] as const,
+    detail: (id: string) => [...keys.base(), storeId, id] as const,
+    newReturns: (lineIds: string[]) =>
+      [...keys.base(), storeId, 'newReturns', lineIds] as const,
+  };
+
+  const { client } = useGql();
+  const queries = getReturnsQueries(getSdk(client), storeId);
+  return { ...queries, storeId, keys };
+};
