+++ client/packages/invoices/src/Returns/api/operations.generated.ts
@@ -1,0 +1,64 @@
+import * as Types from '@openmsupply-client/common';
+
+import { GraphQLClient } from 'graphql-request';
+import { GraphQLClientRequestHeaders } from 'graphql-request/build/cjs/types';
+import gql from 'graphql-tag';
+import { graphql, ResponseResolver, GraphQLRequest, GraphQLContext } from 'msw'
+export type NewSupplierReturnLinesQueryVariables = Types.Exact<{
+  inboundShipmentLineIds?: Types.InputMaybe<Array<Types.Scalars['String']['input']> | Types.Scalars['String']['input']>;
+  storeId: Types.Scalars['String']['input'];
+}>;
+
+
+export type NewSupplierReturnLinesQuery = { __typename: 'Queries', newSupplierReturn: Array<{ __typename: 'SupplierReturnLine', availableNumberOfPacks: number, batch?: string | null, expiryDate?: string | null, id: string, itemCode: string, itemName: string, numberOfPacksToReturn: number, packSize: number, stockLineId: string }> };
+
+
+export const NewSupplierReturnLinesDocument = gql`
+    query newSupplierReturnLines($inboundShipmentLineIds: [String!], $storeId: String!) {
+  newSupplierReturn(
+    input: {inboundShipmentLineIds: $inboundShipmentLineIds}
+    storeId: $storeId
+  ) {
+    availableNumberOfPacks
+    batch
+    expiryDate
+    id
+    itemCode
+    itemName
+    numberOfPacksToReturn
+    packSize
+    stockLineId
+  }
+}
+    `;
+
+export type SdkFunctionWrapper = <T>(action: (requestHeaders?:Record<string, string>) => Promise<T>, operationName: string, operationType?: string) => Promise<T>;
+
+
+const defaultWrapper: SdkFunctionWrapper = (action, _operationName, _operationType) => action();
+
+export function getSdk(client: GraphQLClient, withWrapper: SdkFunctionWrapper = defaultWrapper) {
+  return {
+    newSupplierReturnLines(variables: NewSupplierReturnLinesQueryVariables, requestHeaders?: GraphQLClientRequestHeaders): Promise<NewSupplierReturnLinesQuery> {
+      return withWrapper((wrappedRequestHeaders) => client.request<NewSupplierReturnLinesQuery>(NewSupplierReturnLinesDocument, variables, {...requestHeaders, ...wrappedRequestHeaders}), 'newSupplierReturnLines', 'query');
+    }
+  };
+}
+export type Sdk = ReturnType<typeof getSdk>;
+
+/**
+ * @param resolver a function that accepts a captured request and may return a mocked response.
+ * @see https://mswjs.io/docs/basics/response-resolver
+ * @example
+ * mockNewSupplierReturnLinesQuery((req, res, ctx) => {
+ *   const { inboundShipmentLineIds, storeId } = req.variables;
+ *   return res(
+ *     ctx.data({ newSupplierReturn })
+ *   )
+ * })
+ */
+export const mockNewSupplierReturnLinesQuery = (resolver: ResponseResolver<GraphQLRequest<NewSupplierReturnLinesQueryVariables>, GraphQLContext<NewSupplierReturnLinesQuery>, any>) =>
+  graphql.query<NewSupplierReturnLinesQuery, NewSupplierReturnLinesQueryVariables>(
+    'newSupplierReturnLines',
+    resolver
+  )
