+++ client/packages/invoices/src/Returns/api/operations.graphql
@@ -1,0 +1,19 @@
+query newSupplierReturnLines(
+  $inboundShipmentLineIds: [String!]
+  $storeId: String!
+) {
+  newSupplierReturn(
+    input: { inboundShipmentLineIds: $inboundShipmentLineIds }
+    storeId: $storeId
+  ) {
+    availableNumberOfPacks
+    batch
+    expiryDate
+    id
+    itemCode
+    itemName
+    numberOfPacksToReturn
+    packSize
+    stockLineId
+  }
+}
